requirements.txt
pillow
numpy 
streamlit 
tensorflow
